requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
rdkit
openpyxl
python-docx